:recycle: Separate PhotoAssetGrid
